Fix bugs
Fix ephemeral message not being updated on reset button click.
Fix excludedroles and excludedusers not being taken into account when purging or fetching by role

+++ src/main/java/pl/xnik3e/Guardian/listeners/MessageCommandListener.java
@@ -1,5 +1,7 @@
 package pl.xnik3e.Guardian.listeners;
 
+import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.entities.User;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
@@ -11,6 +13,7 @@ import pl.xnik3e.Guardian.Services.FireStoreService;
 import pl.xnik3e.Guardian.Utils.MessageUtils;
 import pl.xnik3e.Guardian.components.Command.CommandManager;
 
+import java.awt.*;
 import java.util.Objects;
 
 @Component
@@ -47,14 +50,28 @@ public class MessageCommandListener extends ListenerAdapter {
             case "reset":
                 fireStoreService.getModel().getDefaultConfig();
                 fireStoreService.updateConfigModel();
-                event.getMessage().delete().queue();
-                messageUtils.openPrivateChannelAndMessageUser(event.getUser(), "Bot has been reset to factory settings");
+                if (!event.getMessage().isEphemeral()) {
+                    event.getMessage().delete().queue();
+                    messageUtils.openPrivateChannelAndMessageUser(event.getUser(), getMessageEmbed());
+                } else {
+                    event.getHook().editOriginalComponents().queue();
+                    event.getHook().editOriginalEmbeds(getMessageEmbed()).queue();
+                }
                 event.deferEdit().queue();
                 break;
             case "another":
                 break;
         }
 
+    }
+
+    @NotNull
+    private static MessageEmbed getMessageEmbed() {
+        return new EmbedBuilder()
+                .setTitle("Reset complete!")
+                .setDescription("Bot has been reset to factory settings")
+                .setColor(Color.GREEN)
+                .build();
     }
 }
